--- svc/pkg/cluster/util/gen/hash.txt
@@ -1,1 +1,0 @@
-80b6ae8f1c15377d4c11bbb28a5705b982c8ac92
